# l2: add C SIMD implementation