--- a/adapters/repos/db/vector/hnsw/distancer/l2_amd64_test.go
+++ b/adapters/repos/db/vector/hnsw/distancer/l2_amd64_test.go
@@ -32,7 +32,7 @@ func L2PureGo(a, b []float32) float32 {
 }
 
 func Test_L2_DistanceImplementation(t *testing.T) {
-	lengths := []int{1, 4, 16, 31, 32, 35, 64, 67, 128, 130, 256, 260, 384, 390, 768, 777}
+	lengths := []int{1, 4, 16, 20, 31, 32, 35, 64, 67, 128, 130, 256, 260, 384, 390, 768, 777}
 
 	for _, length := range lengths {
 		t.Run(fmt.Sprintf("with vector l=%d", length), func(t *testing.T) {
@@ -52,7 +52,7 @@ func Test_L2_DistanceImplementation(t *testing.T) {
 }
 
 func Test_L2_DistanceImplementation_OneNegativeValue(t *testing.T) {
-	lengths := []int{1, 4, 16, 31, 32, 35, 64, 67, 128, 130, 256, 260, 384, 390, 768, 777}
+	lengths := []int{1, 4, 16, 20, 31, 32, 35, 64, 67, 128, 130, 256, 260, 384, 390, 768, 777}
 
 	for _, length := range lengths {
 		t.Run(fmt.Sprintf("with vector l=%d", length), func(t *testing.T) {
@@ -73,7 +73,7 @@ func Test_L2_DistanceImplementation_OneNegativeValue(t *testing.T) {
 
 func Benchmark_L2_PureGo_VS_AVX(b *testing.B) {
 	r := getRandomSeed()
-	lengths := []int{30, 32, 128, 256, 300, 384, 600, 768, 1024}
+	lengths := []int{2, 4, 6, 8, 10, 12, 16, 24, 30, 32, 64, 128, 256, 300, 384, 600, 768, 1024}
 	for _, length := range lengths {
 		b.Run(fmt.Sprintf("vector dim=%d", length), func(b *testing.B) {
 			x := make([]float32, length)