… adapters/repos/db/vector/hnsw/distancer/asm/l2_amd64.go
… adapters/repos/db/vector/hnsw/distancer/asm/l2_amd64.s
… adapters/repos/db/vector/hnsw/distancer/asm/l2_stub_amd64.go