### What's being changed:
This replaces the current implementation of the L2 distance function with a version written in C and optimized using goat and Clang.

## AVX256 only

```
goos: linux
goarch: amd64
pkg: github.com/weaviate/weaviate/adapters/repos/db/vector/hnsw/distancer
cpu: Intel(R) Xeon(R) CPU @ 2.80GHz
                                    │ l2_asm_current.txt │               v41.txt               │
                                    │       sec/op       │    sec/op     vs base               │
_L2_PureGo_VS_AVX/vector_dim=2-8             5.107n ± 1%   5.585n ±  0%   +9.35% (p=0.002 n=6)
_L2_PureGo_VS_AVX/vector_dim=4-8             6.497n ± 1%   5.333n ±  0%  -17.90% (p=0.002 n=6)
_L2_PureGo_VS_AVX/vector_dim=6-8             8.104n ± 1%   7.945n ±  0%   -1.96% (p=0.002 n=6)
_L2_PureGo_VS_AVX/vector_dim=8-8             9.719n ± 1%   5.832n ±  0%  -39.99% (p=0.002 n=6)
_L2_PureGo_VS_AVX/vector_dim=10-8           11.170n ± 1%   7.569n ±  0%  -32.24% (p=0.002 n=6)
_L2_PureGo_VS_AVX/vector_dim=12-8           12.685n ± 0%   8.678n ±  4%  -31.58% (p=0.002 n=6)
_L2_PureGo_VS_AVX/vector_dim=16-8           15.860n ± 2%   6.577n ±  0%  -58.53% (p=0.002 n=6)
_L2_PureGo_VS_AVX/vector_dim=24-8           23.665n ± 0%   7.321n ±  1%  -69.06% (p=0.002 n=6)
_L2_PureGo_VS_AVX/vector_dim=30-8            29.03n ± 0%   11.85n ±  2%  -59.20% (p=0.002 n=6)
_L2_PureGo_VS_AVX/vector_dim=32-8            7.225n ± 0%   7.899n ±  0%   +9.33% (p=0.002 n=6)
_L2_PureGo_VS_AVX/vector_dim=128-8           12.99n ± 1%   13.87n ±  1%   +6.77% (p=0.002 n=6)
_L2_PureGo_VS_AVX/vector_dim=256-8           21.21n ± 1%   22.57n ±  1%   +6.36% (p=0.002 n=6)
_L2_PureGo_VS_AVX/vector_dim=300-8           35.53n ± 2%   27.09n ± 12%  -23.75% (p=0.002 n=6)
_L2_PureGo_VS_AVX/vector_dim=384-8           31.40n ± 1%   32.98n ±  3%   +5.05% (p=0.002 n=6)
_L2_PureGo_VS_AVX/vector_dim=600-8           66.17n ± 1%   43.02n ±  2%  -34.98% (p=0.002 n=6)
_L2_PureGo_VS_AVX/vector_dim=768-8           49.14n ± 0%   47.70n ±  1%   -2.92% (p=0.002 n=6)
_L2_PureGo_VS_AVX/vector_dim=1024-8          59.91n ± 2%   59.29n ±  0%        ~ (p=0.093 n=6)
_L2_PureGo_VS_AVX/vector_dim=64-8                          9.535n ±  0%
geomean                                      17.76n        13.12n        -24.75%
```

- We can see interesting improvements with dim < 32
- It's slightly slower when 32 <= dim < 256
- Non power of 2 dimensions are much faster in general (ex: dim=300 and dim=600)
- The other ones are equivalent

## Mix AVX256 and AVX512 (for dim > 512)

```
goos: linux
goarch: amd64
pkg: github.com/weaviate/weaviate/adapters/repos/db/vector/hnsw/distancer
cpu: Intel(R) Xeon(R) CPU @ 2.80GHz
                                    │ l2_asm_current.txt │              v43.txt               │
                                    │       sec/op       │   sec/op     vs base               │
_L2_PureGo_VS_AVX/vector_dim=2-8             5.107n ± 1%   5.869n ± 0%  +14.91% (p=0.002 n=6)
_L2_PureGo_VS_AVX/vector_dim=4-8             6.497n ± 1%   5.093n ± 1%  -21.61% (p=0.002 n=6)
_L2_PureGo_VS_AVX/vector_dim=6-8             8.104n ± 1%   6.814n ± 0%  -15.92% (p=0.002 n=6)
_L2_PureGo_VS_AVX/vector_dim=8-8             9.719n ± 1%   7.022n ± 0%  -27.75% (p=0.002 n=6)
_L2_PureGo_VS_AVX/vector_dim=10-8           11.170n ± 1%   9.010n ± 0%  -19.34% (p=0.002 n=6)
_L2_PureGo_VS_AVX/vector_dim=12-8           12.685n ± 0%   8.978n ± 0%  -29.23% (p=0.002 n=6)
_L2_PureGo_VS_AVX/vector_dim=16-8           15.860n ± 2%   7.607n ± 0%  -52.03% (p=0.002 n=6)
_L2_PureGo_VS_AVX/vector_dim=24-8           23.665n ± 0%   8.371n ± 0%  -64.63% (p=0.002 n=6)
_L2_PureGo_VS_AVX/vector_dim=30-8            29.03n ± 0%   11.55n ± 1%  -60.21% (p=0.002 n=6)
_L2_PureGo_VS_AVX/vector_dim=32-8            7.225n ± 0%   8.489n ± 0%  +17.49% (p=0.002 n=6)
_L2_PureGo_VS_AVX/vector_dim=128-8           12.99n ± 1%   13.92n ± 1%   +7.16% (p=0.002 n=6)
_L2_PureGo_VS_AVX/vector_dim=256-8           21.21n ± 1%   22.41n ± 0%   +5.66% (p=0.002 n=6)
_L2_PureGo_VS_AVX/vector_dim=300-8           35.53n ± 2%   28.13n ± 4%  -20.81% (p=0.002 n=6)
_L2_PureGo_VS_AVX/vector_dim=384-8           31.40n ± 1%   32.45n ± 0%   +3.38% (p=0.002 n=6)
_L2_PureGo_VS_AVX/vector_dim=600-8           66.17n ± 1%   46.32n ± 1%  -29.99% (p=0.002 n=6)
_L2_PureGo_VS_AVX/vector_dim=768-8           49.14n ± 0%   45.27n ± 2%   -7.89% (p=0.002 n=6)
_L2_PureGo_VS_AVX/vector_dim=1024-8          59.91n ± 2%   50.10n ± 1%  -16.38% (p=0.002 n=6)
_L2_PureGo_VS_AVX/vector_dim=64-8                          9.821n ± 0%
geomean                                      17.76n        13.47n       -22.77%
```

With AVX512, the results are not really interesting with anything below 600 dimensions, but it gets really interesting with 1024 dimensions (50ns vs 60ns with the current implementation) and above.

### Review checklist

- [ ] Chaos pipeline run or not necessary. Link to pipeline:
- [x] All new code is covered by tests where it is reasonable.
- [x] Performance tests have been run or not necessary.